[refactor] Changed directory structure and fixed test.
