+++ lib/utilities/nativeAdapter.dart
